bbmControl class

bbmControl class is used for setting the control parameters for the bbm function.

bbmControl(object, ...)

# S4 method for missing
bbmControl(object, ...)

Slots

g

Instantaneous rate of biomass decrease, g = M - G, for recruits (rec) and adults (adult), c(rec=numeric(1), adult=numeric(1)).

param.fix

An FLPar with value 1 if the parameter is fixed at its initial value and 0 otherwise.

Validity

Dimensions

g must be a vector of length 2 and with the names 'rec' and 'adult'

Class

param.fix must be of class FLPar

You can inspect the class validity function by using getValidity(getClassDef('bbmControl'))

Accessors

All slots in the class have accessor methods defined that allow retrieving individual slots.

Constructor

A construction method exists for this class that can take named arguments for any of its slots. All slots are then created to match the requirements of the class validity.

See also

Examples

# Load data
data(ane)

# Generate an object of FLPar class (different alternatives)
bbmControl()                     # empty object

#> An object of class "bbmControl"
#> Slot "g":
#>   rec adult 
#>    NA    NA 
#> 
#> Slot "param.fix":
#> An object of class "FLPar"
#> params
#>    
#> NA 
#> units:  NA 
#> 

slotNames(bbmControl())          # slots

#> [1] "g"         "param.fix"

bbmControl( g=c(rec=0.68, adult=0.68), param.fix=FLPar(nyear=20, nindex=3)) # setting values for the slots

#> An object of class "bbmControl"
#> Slot "g":
#>   rec adult 
#>  0.68  0.68 
#> 
#> Slot "param.fix":
#> An object of class "FLPar"
#> params
#>  nyear nindex 
#>     20      3 
#> units:  NA 
#> 

# Run assessment (control must be of class bbmControl)
class(control.ane)

#> [1] "bbmControl"
#> attr(,"package")
#> [1] "bbm"

run <- bbm(catch.ane, indicesB=indicesB.ane, indicesP=indicesP.ane, control=control.ane, inits=inits.ane)

#> outer mgc:  2.991218 
#> outer mgc:  2.983554 
#> outer mgc:  2.994433 
#> outer mgc:  2.980339 
#> outer mgc:  2.989589 
#> outer mgc:  2.985185 
#> outer mgc:  2.990111 
#> outer mgc:  2.984663 
#> outer mgc:  2.986485 
#> outer mgc:  2.988286 
#> outer mgc:  2.987929 
#> outer mgc:  2.986843 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987385 
#> outer mgc:  2.987387 
#> outer mgc:  2.987385 
#> outer mgc:  2.987387 
#> outer mgc:  2.987382 
#> outer mgc:  2.98739 
#> outer mgc:  2.987383 
#> outer mgc:  2.987388 
#> outer mgc:  2.98738 
#> outer mgc:  2.987392 
#> outer mgc:  2.987369 
#> outer mgc:  2.987403 
#> outer mgc:  2.987363 
#> outer mgc:  2.987409 
#> outer mgc:  2.987322 
#> outer mgc:  2.98745 
#> outer mgc:  2.98723 
#> outer mgc:  2.987541 
#> outer mgc:  2.987124 
#> outer mgc:  2.987647 
#> outer mgc:  2.986773 
#> outer mgc:  2.987998 
#> outer mgc:  2.986082 
#> outer mgc:  2.988689 
#> outer mgc:  2.986608 
#> outer mgc:  2.988164 
#> outer mgc:  3.008474 
#> outer mgc:  2.96631 
#> outer mgc:  2.986905 
#> outer mgc:  2.987863 
#> outer mgc:  2.986374 
#> outer mgc:  2.988397 
#> outer mgc:  2.986725 
#> outer mgc:  2.988047 
#> outer mgc:  2.9823 
#> outer mgc:  2.992471 
#> outer mgc:  2.985849 
#> outer mgc:  2.988921 
#> outer mgc:  2.987386 
#> outer mgc:  2.991218 
#> outer mgc:  2.983554 
#> outer mgc:  2.994433 
#> outer mgc:  2.980339 
#> outer mgc:  2.989589 
#> outer mgc:  2.985185 
#> outer mgc:  2.990111 
#> outer mgc:  2.984663 
#> outer mgc:  2.986485 
#> outer mgc:  2.988286 
#> outer mgc:  2.987929 
#> outer mgc:  2.986843 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987386 
#> outer mgc:  2.987385 
#> outer mgc:  2.987387 
#> outer mgc:  2.987385 
#> outer mgc:  2.987387 
#> outer mgc:  2.987382 
#> outer mgc:  2.98739 
#> outer mgc:  2.987383 
#> outer mgc:  2.987388 
#> outer mgc:  2.98738 
#> outer mgc:  2.987392 
#> outer mgc:  2.987369 
#> outer mgc:  2.987403 
#> outer mgc:  2.987363 
#> outer mgc:  2.987409 
#> outer mgc:  2.987322 
#> outer mgc:  2.98745 
#> outer mgc:  2.98723 
#> outer mgc:  2.987541 
#> outer mgc:  2.987124 
#> outer mgc:  2.987647 
#> outer mgc:  2.986773 
#> outer mgc:  2.987998 
#> outer mgc:  2.986082 
#> outer mgc:  2.988689 
#> outer mgc:  2.986608 
#> outer mgc:  2.988164 
#> outer mgc:  3.008474 
#> outer mgc:  2.96631 
#> outer mgc:  2.986905 
#> outer mgc:  2.987863 
#> outer mgc:  2.986374 
#> outer mgc:  2.988397 
#> outer mgc:  2.986725 
#> outer mgc:  2.988047 
#> outer mgc:  2.9823 
#> outer mgc:  2.992471 
#> outer mgc:  2.985849 
#> outer mgc:  2.988921 
#> outer mgc:  282597.9 

#> Warning: NaNs produced

#> Warning: NaNs produced

run

#> An object of class "bbmFit"
#> Slot "inputs":
#> $catch
#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>    year
#> age 1987       1988       1989       1990       1991       1992      
#>   1  2711.0000  2602.0000  1723.0000  9314.0000  3903.0000 11933.0000
#>   2  5607.0000  1262.0000  2153.0000  1259.0000  6288.0000  4433.0000
#>    year
#> age 1993       1994       1995       1996       1997       1998      
#>   1  6414.0000  3795.0000  5718.0000  4570.0000  4323.0000  5898.0000
#>   2  7763.0000  9807.0000  8832.0000  4676.0000  2912.0000  2090.0000
#>    year
#> age 1999       2000       2001       2002       2003       2004      
#>   1  2067.0000  6298.0000  5481.0000  1962.0000   625.0000  2754.0000
#>   2  8828.0000  5712.0000  5987.0000  5776.0000  1754.0000  1869.0000
#>    year
#> age 2005       2006      
#>   1   102.0000   287.0000
#>   2   688.0000   311.0000
#> 
#> , , unit = unique, season = 2, area = unique
#> 
#>    year
#> age 1987       1988       1989       1990       1991       1992      
#>   1  3960.0000  8554.0000  2039.0000 20249.0000  5048.0000 17565.0000
#>   2  2583.0000  2400.0000  2403.0000  3325.0000  3148.0000  3461.0000
#>    year
#> age 1993       1994       1995       1996       1997       1998      
#>   1 17812.0000 12425.0000  9560.0000 17869.0000 11450.0000 17841.0000
#>   2  7619.0000  7725.0000  5255.0000  5964.0000  1806.0000  5747.0000
#>    year
#> age 1999       2000       2001       2002       2003       2004      
#>   1  7753.0000 16957.0000 17911.0000  4367.0000  3848.0000  9211.0000
#>   2  7758.0000  7925.0000 10760.0000  5387.0000  4253.0000  2446.0000
#>    year
#> age 2005       2006      
#>   1    81.0000     0.0005
#>   2   291.0000     0.0005
#> 
#> units:  NA 
#> 
#> $indicesB
#> $ depm 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all  29365  63500  16720  97239  19276  90720     NA  60062  54700  39545
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all  51176 101976  69074  44973 120403  30697  23962  19498   8002  21436
#> 
#> units:  NA 
#> 
#> $ acoustic 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   all     NA     NA     NA     NA  64000  89000     NA  35000     NA     NA
#>      year
#> age   1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   all  63000  57000     NA  98484 137200  97051  29430  46018  15603  30649
#> 
#> units:  NA 
#> 
#> 
#> $indicesP
#> $ depm 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1987    1988    1989    1990    1991    1992    1993    1994    1995   
#>   all 0.48476 0.83602 0.43553 0.93224 0.58472 0.94324      NA 0.57730 0.78439
#>      year
#> age   1996    1997    1998    1999    2000    2001    2002    2003    2004   
#>   all      NA 0.75301 0.78800      NA      NA 0.57399 0.20693 0.69172 0.75131
#>      year
#> age   2005    2006   
#>   all 0.25781 0.71282
#> 
#> units:  NA 
#> 
#> $ acoustic 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1987    1988    1989    1990    1991    1992    1993    1994    1995   
#>   all      NA      NA      NA      NA 0.44253 0.94875      NA      NA      NA
#>      year
#> age   1996    1997    1998    1999    2000    2001    2002    2003    2004   
#>   all      NA 0.61108      NA      NA      NA 0.66274 0.18262 0.53456 0.80673
#>      year
#> age   2005    2006   
#>   all 0.15414 0.54442
#> 
#> units:  NA 
#> 
#> 
#> $perindicesB
#>     depm acoustic 
#>        2        2 
#> 
#> $perindicesP
#>     depm acoustic 
#>        2        2 
#> 
#> $control
#> An object of class "bbmControl"
#> Slot "g":
#>   rec adult 
#>  0.68  0.68 
#> 
#> Slot "param.fix":
#> An object of class "FLPar"
#> params
#>       q_depm   q_acoustic     psi_depm psi_acoustic      xi_depm  xi_acoustic 
#>            0            0            0            0            0            0 
#>           B0       R_1987       R_1988       R_1989       R_1990       R_1991 
#>            0            0            0            0            0            0 
#>       R_1992       R_1993       R_1994       R_1995       R_1996       R_1997 
#>            0            0            0            0            0            0 
#>       R_1998       R_1999       R_2000       R_2001       R_2002       R_2003 
#>            0            0            0            0            0            0 
#>       R_2004       R_2005       R_2006          mur         psir 
#>            0            0            0            0            0 
#> units:  NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA 
#> 
#> 
#> $f
#> [1] 0.375 0.625
#> 
#> $nper
#> [1] 2
#> 
#> 
#> Slot "convergence":
#> [1] 0
#> 
#> Slot "message":
#> [1] ""
#> 
#> Slot "fitSumm":
#>        iter
#>                1
#>   nlogl 15.22352
#>   nobs  31.00000
#>   nopar 29.00000
#> 
#> Slot "params":
#> An object of class "FLPar"
#> params
#>       q_depm   q_acoustic     psi_depm psi_acoustic      xi_depm  xi_acoustic 
#>     2.64e-01     3.44e-01     3.90e+00     7.17e+00     2.24e+00     2.38e+00 
#>           B0       R_1987       R_1988       R_1989       R_1990       R_1991 
#>     6.55e+04     8.91e+04     1.27e+05     7.10e+04     1.69e+05     8.83e+04 
#>       R_1992       R_1993       R_1994       R_1995       R_1996       R_1997 
#>     2.83e+05     8.70e+04     1.01e+05     1.56e+05     1.05e+05     1.48e+05 
#>       R_1998       R_1999       R_2000       R_2001       R_2002       R_2003 
#>     1.82e+05     1.55e+05     1.83e+05     1.98e+05     5.98e+04     8.43e+04 
#>       R_2004       R_2005       R_2006          mur         psir 
#>     1.14e+05     3.61e+04     9.31e+04     1.16e+01     5.09e+00 
#> units:  NA 
#> 
#> Slot "params.se":
#> An object of class "FLPar"
#> params
#>       q_depm   q_acoustic     psi_depm psi_acoustic      xi_depm  xi_acoustic 
#>     3.82e-02     5.12e-02     1.43e+00     3.57e+00     2.72e-01     5.67e-01 
#>           B0       R_1987       R_1988       R_1989       R_1990       R_1991 
#>     2.56e+04     2.81e+03     3.39e+04          NaN     5.21e+04     1.30e+04 
#>       R_1992       R_1993       R_1994       R_1995       R_1996       R_1997 
#>     1.19e+05     2.38e+04     2.55e+04     6.37e+04     3.02e+04     4.19e+04 
#>       R_1998       R_1999       R_2000       R_2001       R_2002       R_2003 
#>     7.39e+04     6.88e+04     6.03e+04     5.24e+04          NaN          NaN 
#>       R_2004       R_2005       R_2006          mur         psir 
#>     2.85e+04          NaN     1.11e+04          NaN     1.99e+00 
#> units:  NA 
#> 
#> Slot "vcov":
#> , , 1
#> 
#>                    logq_depm logq_acoustic logpsi_depm logpsi_acoustic
#> logq_depm        0.020957252   0.007664286 -0.07639773    -0.055644319
#> logq_acoustic    0.007664286   0.022113190 -0.06971259    -0.058356490
#> logpsi_depm     -0.076397727  -0.069712587  0.13521670     0.015470203
#> logpsi_acoustic -0.055644319  -0.058356490  0.01547020     0.247261444
#> xi_depm         -0.093958781  -0.099275446 -0.04250018    -0.070744687
#> xi_acoustic     -0.142554296  -0.146216587 -0.02675519    -0.075436817
#> logB0            0.005807806   0.014334935  0.06944297     0.059293551
#> logR_1987        0.043823506   0.050461191  0.07535188     0.068811933
#> logR_1988       -0.007619600  -0.003262914  0.06732703     0.053285285
#> logR_1989        0.065236569   0.068932606  0.06043592     0.066764179
#> logR_1990       -0.035467860  -0.036230759  0.06251330     0.053733353
#> logR_1991        0.018938185   0.016923913  0.06272157     0.073226588
#> logR_1992       -0.077312929  -0.080237708  0.06665922     0.031036625
#> logR_1993        0.067735788   0.069213018  0.05047386     0.023555438
#> logR_1994        0.004080910   0.004307739  0.06967180     0.025041134
#> logR_1995       -0.048987939  -0.045666943  0.07586300     0.044614804
#> logR_1996        0.026073838   0.026626086  0.04682051     0.044048729
#> logR_1997       -0.017385516  -0.019422135  0.07834092     0.056123901
#> logR_1998       -0.050282010  -0.052188195  0.07882731     0.023991406
#> logR_1999       -0.022366019  -0.017839692  0.06378648     0.056764977
#> logR_2000       -0.012224509  -0.018591861  0.03418522     0.070661442
#> logR_2001       -0.021666320  -0.025123300  0.05627042     0.065352122
#> logR_2002        0.076959536   0.075089343  0.05975568     0.074319659
#> logR_2003        0.031807184   0.028168725  0.04589560     0.016350700
#> logR_2004       -0.001262388  -0.006134466  0.03148665    -0.004674095
#> logR_2005        0.116688536   0.114477326  0.04447625     0.018534233
#> logR_2006        0.028683333   0.025167750  0.05085288     0.021652400
#> mur              0.009409654   0.008871890  0.05907433     0.044152698
#> logpsir          0.194754560   0.196693406 -0.01626422    -0.010008751
#>                     xi_depm xi_acoustic        logB0     logR_1987    logR_1988
#> logq_depm       -0.09395878 -0.14255430  0.005807806  0.0438235058 -0.007619600
#> logq_acoustic   -0.09927545 -0.14621659  0.014334935  0.0504611906 -0.003262914
#> logpsi_depm     -0.04250018 -0.02675519  0.069442973  0.0753518814  0.067327030
#> logpsi_acoustic -0.07074469 -0.07543682  0.059293551  0.0688119333  0.053285285
#> xi_depm          0.07390166 -0.04201099  0.067337114  0.0696913629  0.078428157
#> xi_acoustic     -0.04201099  0.32118232  0.116128479  0.1284274286  0.114087282
#> logB0            0.06733711  0.11612848  0.153262059 -0.0394739361 -0.006849633
#> logR_1987        0.06969136  0.12842743 -0.039473936  0.0009938772 -0.038210219
#> logR_1988        0.07842816  0.11408728 -0.006849633 -0.0382102189  0.071576643
#> logR_1989        0.08460310  0.12209549 -0.064891476 -0.0987260482 -0.051211570
#> logR_1990        0.06891231  0.08361449  0.012689985 -0.0132813378  0.026036649
#> logR_1991        0.09054156  0.08770601 -0.032323822 -0.0653786776 -0.016900543
#> logR_1992        0.06868115  0.16132813  0.046650019  0.0179796427  0.062562406
#> logR_1993        0.07407151  0.07975942 -0.069627275 -0.0909559457 -0.060031457
#> logR_1994        0.08103983  0.11822797 -0.020875874 -0.0494878118 -0.007468251
#> logR_1995        0.07473276  0.12890012  0.022047892 -0.0080740425  0.037265540
#> logR_1996        0.05387489  0.09413956 -0.033119587 -0.0540669223 -0.024655763
#> logR_1997        0.08449539  0.12583016 -0.004523135 -0.0373761314  0.011245153
#> logR_1998        0.08851098  0.12813307  0.021345131 -0.0099455955  0.038527088
#> logR_1999        0.05994561  0.10017432  0.004058793 -0.0208462590  0.016178629
#> logR_2000        0.05329742  0.07761005 -0.001702424 -0.0238923569  0.008613997
#> logR_2001        0.06036118  0.11898887  0.003377632 -0.0227784139  0.015343114
#> logR_2002        0.10058334  0.08573877 -0.081580042 -0.1159707540 -0.067747705
#> logR_2003        0.11746987  0.13422141 -0.049175348 -0.0848356044 -0.031775832
#> logR_2004        0.10902656  0.19444034 -0.022331535 -0.0552492068 -0.005234377
#> logR_2005        0.14478519  0.11988966 -0.120827706 -0.1599571001 -0.103549456
#> logR_2006        0.12193066  0.13447253 -0.047159486 -0.0833212224 -0.028636914
#> mur              0.08429733  0.11687709 -0.024330182 -0.0507868522 -0.007521225
#> logpsir          0.06848135 -0.01728292 -0.164272879 -0.1674772853 -0.167102532
#>                   logR_1989    logR_1990    logR_1991   logR_1992   logR_1993
#> logq_depm        0.06523657 -0.035467860  0.018938185 -0.07731293  0.06773579
#> logq_acoustic    0.06893261 -0.036230759  0.016923913 -0.08023771  0.06921302
#> logpsi_depm      0.06043592  0.062513300  0.062721565  0.06665922  0.05047386
#> logpsi_acoustic  0.06676418  0.053733353  0.073226588  0.03103662  0.02355544
#> xi_depm          0.08460310  0.068912312  0.090541564  0.06868115  0.07407151
#> xi_acoustic      0.12209549  0.083614491  0.087706012  0.16132813  0.07975942
#> logB0           -0.06489148  0.012689985 -0.032323822  0.04665002 -0.06962728
#> logR_1987       -0.09872605 -0.013281338 -0.065378678  0.01797964 -0.09095595
#> logR_1988       -0.05121157  0.026036649 -0.016900543  0.06256241 -0.06003146
#> logR_1989       -0.03251213 -0.030554124 -0.079657913 -0.00176436 -0.10067066
#> logR_1990       -0.03055412  0.095645945  0.017960022  0.08577658 -0.04163747
#> logR_1991       -0.07965791  0.017960022  0.021738348  0.04470045 -0.08135101
#> logR_1992       -0.00176436  0.085776577  0.044700445  0.17731210 -0.03923294
#> logR_1993       -0.10067066 -0.041637467 -0.081351009 -0.03923294  0.07499174
#> logR_1994       -0.06605587  0.016127565 -0.029750353  0.04794345 -0.06432912
#> logR_1995       -0.02883913  0.060713506  0.015396512  0.10542621 -0.03641153
#> logR_1996       -0.06609205 -0.005989840 -0.041701417  0.01357802 -0.06429205
#> logR_1997       -0.05745195  0.039768263 -0.010306190  0.07954903 -0.05924819
#> logR_1998       -0.02971375  0.062875914  0.016784305  0.10794985 -0.03700232
#> logR_1999       -0.03804797  0.036892018 -0.002198503  0.06808998 -0.04402938
#> logR_2000       -0.03559990  0.029486558 -0.001381485  0.05353288 -0.04352667
#> logR_2001       -0.03949172  0.037889927 -0.001947956  0.07334805 -0.04763288
#> logR_2002       -0.12908112 -0.032612148 -0.085614081 -0.01556720 -0.10539396
#> logR_2003       -0.09806737 -0.001228359 -0.054183465  0.03419634 -0.08213535
#> logR_2004       -0.06894988  0.018591718 -0.031880133  0.06972287 -0.06333933
#> logR_2005       -0.16825700 -0.066395908 -0.126761185 -0.04634233 -0.12448016
#> logR_2006       -0.09690308  0.002551320 -0.051270430  0.03824531 -0.08121200
#> mur             -0.06598735  0.016818987 -0.028302483  0.04871177 -0.05965318
#> logpsir         -0.14918662 -0.158961113 -0.166697729 -0.19689254 -0.08120639
#>                    logR_1994    logR_1995    logR_1996     logR_1997
#> logq_depm        0.004080910 -0.048987939  0.026073838 -0.0173855162
#> logq_acoustic    0.004307739 -0.045666943  0.026626086 -0.0194221350
#> logpsi_depm      0.069671797  0.075862999  0.046820514  0.0783409171
#> logpsi_acoustic  0.025041134  0.044614804  0.044048729  0.0561239007
#> xi_depm          0.081039833  0.074732764  0.053874894  0.0844953943
#> xi_acoustic      0.118227974  0.128900120  0.094139563  0.1258301574
#> logB0           -0.020875874  0.022047892 -0.033119587 -0.0045231350
#> logR_1987       -0.049487812 -0.008074042 -0.054066922 -0.0373761314
#> logR_1988       -0.007468251  0.037265540 -0.024655763  0.0112451532
#> logR_1989       -0.066055868 -0.028839126 -0.066092052 -0.0574519452
#> logR_1990        0.016127565  0.060713506 -0.005989840  0.0397682630
#> logR_1991       -0.029750353  0.015396512 -0.041701417 -0.0103061901
#> logR_1992        0.047943453  0.105426210  0.013578021  0.0795490318
#> logR_1993       -0.064329120 -0.036411534 -0.064292052 -0.0592481938
#> logR_1994        0.063877977  0.031426448 -0.039303953  0.0024011180
#> logR_1995        0.031426448  0.167149423 -0.034842403  0.0535762430
#> logR_1996       -0.039303953 -0.034842403  0.082081803 -0.0151420159
#> logR_1997        0.002401118  0.053576243 -0.015142016  0.0805885369
#> logR_1998        0.030899609  0.080070730 -0.001868924  0.0544252249
#> logR_1999        0.005107669  0.046446574 -0.014895217  0.0200736540
#> logR_2000       -0.002147194  0.033300083 -0.014231346  0.0150393179
#> logR_2001        0.004979568  0.048267460 -0.011389069  0.0266809694
#> logR_2002       -0.074268805 -0.038746947 -0.073906162 -0.0655265890
#> logR_2003       -0.035264666  0.003861953 -0.050561755 -0.0244525383
#> logR_2004       -0.008126048  0.033008326 -0.029091709  0.0046943079
#> logR_2005       -0.101906885 -0.074730976 -0.100557011 -0.1055193636
#> logR_2006       -0.032849313  0.007941800 -0.049285185 -0.0209387319
#> mur             -0.015515475  0.024975936 -0.029887265 -0.0005302686
#> logpsir         -0.150261377 -0.190935930 -0.108884421 -0.1908252530
#>                    logR_1998    logR_1999     logR_2000    logR_2001
#> logq_depm       -0.050282010 -0.022366019 -0.0122245085 -0.021666320
#> logq_acoustic   -0.052188195 -0.017839692 -0.0185918605 -0.025123300
#> logpsi_depm      0.078827313  0.063786480  0.0341852220  0.056270423
#> logpsi_acoustic  0.023991406  0.056764977  0.0706614424  0.065352122
#> xi_depm          0.088510985  0.059945606  0.0532974160  0.060361176
#> xi_acoustic      0.128133074  0.100174321  0.0776100546  0.118988874
#> logB0            0.021345131  0.004058793 -0.0017024241  0.003377632
#> logR_1987       -0.009945596 -0.020846259 -0.0238923569 -0.022778414
#> logR_1988        0.038527088  0.016178629  0.0086139974  0.015343114
#> logR_1989       -0.029713750 -0.038047969 -0.0355998978 -0.039491716
#> logR_1990        0.062875914  0.036892018  0.0294865578  0.037889927
#> logR_1991        0.016784305 -0.002198503 -0.0013814852 -0.001947956
#> logR_1992        0.107949848  0.068089985  0.0535328820  0.073348045
#> logR_1993       -0.037002319 -0.044029376 -0.0435266659 -0.047632883
#> logR_1994        0.030899609  0.005107669 -0.0021471937  0.004979568
#> logR_1995        0.080070730  0.046446574  0.0333000834  0.048267460
#> logR_1996       -0.001868924 -0.014895217 -0.0142313458 -0.011389069
#> logR_1997        0.054425225  0.020073654  0.0150393179  0.026680969
#> logR_1998        0.165474142  0.030618719  0.0243272217  0.048372950
#> logR_1999        0.030618719  0.198289668 -0.0331683462  0.024405032
#> logR_2000        0.024327222 -0.033168346  0.1080934566  0.013487362
#> logR_2001        0.048372950  0.024405032  0.0134873622  0.070264069
#> logR_2002       -0.036946119 -0.046232533 -0.0378516979 -0.046191603
#> logR_2003        0.011033541 -0.016237495 -0.0150142558 -0.013372027
#> logR_2004        0.040199467  0.006311785  0.0030209418  0.012039418
#> logR_2005       -0.066833113 -0.081324731 -0.0709496851 -0.082662291
#> logR_2006        0.015215678 -0.013033741 -0.0133212121 -0.011622538
#> mur              0.027091758  0.007009931 -0.0002021729  0.004785408
#> logpsir         -0.186844253 -0.156394163 -0.1322482694 -0.162101376
#>                   logR_2002    logR_2003    logR_2004   logR_2005   logR_2006
#> logq_depm        0.07695954  0.031807184 -0.001262388  0.11668854  0.02868333
#> logq_acoustic    0.07508934  0.028168725 -0.006134466  0.11447733  0.02516775
#> logpsi_depm      0.05975568  0.045895599  0.031486650  0.04447625  0.05085288
#> logpsi_acoustic  0.07431966  0.016350700 -0.004674095  0.01853423  0.02165240
#> xi_depm          0.10058334  0.117469874  0.109026563  0.14478519  0.12193066
#> xi_acoustic      0.08573877  0.134221412  0.194440342  0.11988966  0.13447253
#> logB0           -0.08158004 -0.049175348 -0.022331535 -0.12082771 -0.04715949
#> logR_1987       -0.11597075 -0.084835604 -0.055249207 -0.15995710 -0.08332122
#> logR_1988       -0.06774770 -0.031775832 -0.005234377 -0.10354946 -0.02863691
#> logR_1989       -0.12908112 -0.098067365 -0.068949879 -0.16825700 -0.09690308
#> logR_1990       -0.03261215 -0.001228359  0.018591718 -0.06639591  0.00255132
#> logR_1991       -0.08561408 -0.054183465 -0.031880133 -0.12676118 -0.05127043
#> logR_1992       -0.01556720  0.034196344  0.069722865 -0.04634233  0.03824531
#> logR_1993       -0.10539396 -0.082135352 -0.063339332 -0.12448016 -0.08121200
#> logR_1994       -0.07426881 -0.035264666 -0.008126048 -0.10190689 -0.03284931
#> logR_1995       -0.03874695  0.003861953  0.033008326 -0.07473098  0.00794180
#> logR_1996       -0.07390616 -0.050561755 -0.029091709 -0.10055701 -0.04928519
#> logR_1997       -0.06552659 -0.024452538  0.004694308 -0.10551936 -0.02093873
#> logR_1998       -0.03694612  0.011033541  0.040199467 -0.06683311  0.01521568
#> logR_1999       -0.04623253 -0.016237495  0.006311785 -0.08132473 -0.01303374
#> logR_2000       -0.03785170 -0.015014256  0.003020942 -0.07094969 -0.01332121
#> logR_2001       -0.04619160 -0.013372027  0.012039418 -0.08266229 -0.01162254
#> logR_2002       -0.06933781 -0.102576488 -0.081845340 -0.17185499 -0.10239002
#> logR_2003       -0.10257649 -0.007355274 -0.020473249 -0.12527270 -0.05110555
#> logR_2004       -0.08184534 -0.020473249  0.062756900 -0.09010667 -0.01535416
#> logR_2005       -0.17185499 -0.125272702 -0.090106668 -0.12542760 -0.12283782
#> logR_2006       -0.10239002 -0.051105550 -0.015354163 -0.12283782  0.01429993
#> mur             -0.07508541 -0.038353132 -0.011079852 -0.10587167 -0.03490579
#> logpsir         -0.14539792 -0.159347223 -0.162912033 -0.12125696 -0.16257724
#>                           mur     logpsir
#> logq_depm        0.0094096542  0.19475456
#> logq_acoustic    0.0088718899  0.19669341
#> logpsi_depm      0.0590743306 -0.01626422
#> logpsi_acoustic  0.0441526980 -0.01000875
#> xi_depm          0.0842973295  0.06848135
#> xi_acoustic      0.1168770940 -0.01728292
#> logB0           -0.0243301823 -0.16427288
#> logR_1987       -0.0507868522 -0.16747729
#> logR_1988       -0.0075212255 -0.16710253
#> logR_1989       -0.0659873541 -0.14918662
#> logR_1990        0.0168189865 -0.15896111
#> logR_1991       -0.0283024826 -0.16669773
#> logR_1992        0.0487117731 -0.19689254
#> logR_1993       -0.0596531752 -0.08120639
#> logR_1994       -0.0155154748 -0.15026138
#> logR_1995        0.0249759365 -0.19093593
#> logR_1996       -0.0298872654 -0.10888442
#> logR_1997       -0.0005302686 -0.19082525
#> logR_1998        0.0270917575 -0.18684425
#> logR_1999        0.0070099310 -0.15639416
#> logR_2000       -0.0002021729 -0.13224827
#> logR_2001        0.0047854079 -0.16210138
#> logR_2002       -0.0750854127 -0.14539792
#> logR_2003       -0.0383531320 -0.15934722
#> logR_2004       -0.0110798521 -0.16291203
#> logR_2005       -0.1058716714 -0.12125696
#> logR_2006       -0.0349057921 -0.16257724
#> mur             -0.0099920466 -0.15576797
#> logpsir         -0.1557679730  0.15285706
#> 
#> 
#> Slot "stock.bio":
#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>    year
#> age 1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   1  89146 126704  70987 168592  88314 282598  86979 100859 155875 105442
#>   2  65519  68279  87701  74572  98045  81921 158252  95516  75366  96798
#>    year
#> age 1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   1 147619 181569 154609 183396 197752  59815  84329 113809  36061  93076
#>   2  77867  99353 118650 119626 126486 134482  86094  78420  85301  60728
#> 
#> , , unit = unique, season = 2, area = unique
#> 
#>    year
#> age 1987   1988   1989   1990   1991   1992   1993   1994   1995   1996  
#>   1  66694  95894  53492 122446  65000 208485  61755  74816 115757  77686
#>   2  45836  51799  66065  56678  70442  59579 115798  65384  50628  70894
#>    year
#> age 1997   1998   1999   2000   2001   2002   2003   2004   2005   2006  
#>   1 110587 135509 117989 136572 148416  44624  64798  85768  27855  71873
#>   2  57777  75150  84173  87672  92745  99128  65172  59124  65495  46786
#> 
#> units:  NA 
#> 
#> Slot "indicesB":
#> $ depm 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1987  1988  1989  1990  1991  1992  1993  1994  1995  1996  1997  1998 
#>   all 29731 39021 31587 47325 35784 70823 46910 37041 43959 39255 44482 55657
#>      year
#> age   1999  2000  2001  2002  2003  2004  2005  2006 
#>   all 53412 59246 63715 37980 34338 38281 24663 31350
#> 
#> units:  NA 
#> 
#> $ acoustic 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1987  1988  1989  1990  1991  1992  1993  1994  1995  1996  1997  1998 
#>   all 38754 50864 41174 61688 46644 92318 61147 48283 57301 51169 57982 72548
#>      year
#> age   1999  2000  2001  2002  2003  2004  2005  2006 
#>   all 69622 77227 83053 49506 44760 49899 32148 40865
#> 
#> units:  NA 
#> 
#> 
#> Slot "indicesP":
#> $ depm 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1987    1988    1989    1990    1991    1992    1993    1994    1995   
#>   all 0.59268 0.64928 0.44742 0.68358 0.47991 0.77774 0.34781 0.53364 0.69572
#>      year
#> age   1996    1997    1998    1999    2000    2001    2002    2003    2004   
#>   all 0.52285 0.65683 0.64326 0.58364 0.60903 0.61542 0.31043 0.49856 0.59195
#>      year
#> age   2005    2006   
#>   all 0.29839 0.60571
#> 
#> units:  NA 
#> 
#> $ acoustic 
#> An object of class "FLQuant"
#> , , unit = unique, season = all, area = unique
#> 
#>      year
#> age   1987    1988    1989    1990    1991    1992    1993    1994    1995   
#>   all 0.59268 0.64928 0.44742 0.68358 0.47991 0.77774 0.34781 0.53364 0.69572
#>      year
#> age   1996    1997    1998    1999    2000    2001    2002    2003    2004   
#>   all 0.52285 0.65683 0.64326 0.58364 0.60903 0.61542 0.31043 0.49856 0.59195
#>      year
#> age   2005    2006   
#>   all 0.29839 0.60571
#> 
#> units:  NA 
#> 
#>