Function to estimate fitted indices

This function estimates fitted indices, given information on growth, periods duration, catches and estimated parameters.

calcIndices(g, findicesB, findicesP, catch, inits)

Arguments

g
A vector with information on instantaneous rate of biomass decrease, g = M - G, for recruits (rec) and adults (adult).

findicesB
A vector with fraction of the year corresponding to each of the indicesB.

findicesP
A vector with fraction of the year corresponding to each of the indicesP.

catch
An FLQuant with catch information (for recruits and adults).

inits
An FLPar with parameter values for the parameters required by the bbm function.

Value

A list with information on estimated indices (FLQuants). The list has 2 elements: indicesB (for indices in biomass) and indicesP (for proportions of recruits in biomass).

Methods

Methods exist for various calculations based on the output class (FLPar). For details: ?FLPar.

See also

Examples

# Load data
data(ane)

args(calcIndices)

#> function (g, findicesB, findicesP, catch, inits) 
#> NULL

indices <- calcIndices( g=control.ane@g,
                        findicesB=unlist(lapply( indicesB.ane, function(x) mean(range(x)[c('startf','endf')]))),
                        findicesP=unlist(lapply( indicesP.ane, function(x) mean(range(x)[c('startf','endf')]))),
                        catch=catch.ane, inits=inits.ane )

class(indices)

#> [1] "list"

slotNames(indices)

#> NULL