bbm package datasets

Data of Bay of Biscay anchovy from Ibaibarriaga et al. (2008).

Example dataset for the classes defined in bbm package. At the moment there is only one dataset of Bay of Biscay anchvoy (ICES Subarea 8), with information on catches together with surveys observations and timing. These are the same data as used in Ibaibarriaga et al. (2008).

Dataset can be loaded by issuing the data command, like in data(ane). and has defined the following objects:

catch.ane, FLQuant Catch information in biomass for recruits (age 1 individuals) and adults.

indicesB.ane, FLIndices Observations from surveys of total biomass, along with survey timing.

indicesP.ane, FLIndices Observations from surveys of proportions of recruits in biomass, along with survey timing.

control.ane, bbmControl Object with information on instantaneous rate of biomass decrease (g = M - G, vector) and information on whether bbm parameters are fixed or not (param.fix, FLPar).

inits.ane, FLPar Initial values for the parameters of the bbm assessment model.

References

Ibaibarriaga et al. (2008). "A Two-Stage Biomass Dynamic Model for Bay of Biscay Anchovy: A Bayesian Approach." ICES J. of Mar. Sci. 65: 191-205.

See also

Examples

data(ane)
ls()

#> character(0)

is(catch.ane)

#> [1] "FLQuant"   "FLArray"   "array"     "structure" "vector"   

catch.ane

#> An object of class "FLQuant"
#> , , unit = unique, season = 1, area = unique
#> 
#>    year
#> age 1987       1988       1989       1990       1991       1992      
#>   1  2711.0000  2602.0000  1723.0000  9314.0000  3903.0000 11933.0000
#>   2  5607.0000  1262.0000  2153.0000  1259.0000  6288.0000  4433.0000
#>    year
#> age 1993       1994       1995       1996       1997       1998      
#>   1  6414.0000  3795.0000  5718.0000  4570.0000  4323.0000  5898.0000
#>   2  7763.0000  9807.0000  8832.0000  4676.0000  2912.0000  2090.0000
#>    year
#> age 1999       2000       2001       2002       2003       2004      
#>   1  2067.0000  6298.0000  5481.0000  1962.0000   625.0000  2754.0000
#>   2  8828.0000  5712.0000  5987.0000  5776.0000  1754.0000  1869.0000
#>    year
#> age 2005       2006      
#>   1   102.0000   287.0000
#>   2   688.0000   311.0000
#> 
#> , , unit = unique, season = 2, area = unique
#> 
#>    year
#> age 1987       1988       1989       1990       1991       1992      
#>   1  3960.0000  8554.0000  2039.0000 20249.0000  5048.0000 17565.0000
#>   2  2583.0000  2400.0000  2403.0000  3325.0000  3148.0000  3461.0000
#>    year
#> age 1993       1994       1995       1996       1997       1998      
#>   1 17812.0000 12425.0000  9560.0000 17869.0000 11450.0000 17841.0000
#>   2  7619.0000  7725.0000  5255.0000  5964.0000  1806.0000  5747.0000
#>    year
#> age 1999       2000       2001       2002       2003       2004      
#>   1  7753.0000 16957.0000 17911.0000  4367.0000  3848.0000  9211.0000
#>   2  7758.0000  7925.0000 10760.0000  5387.0000  4253.0000  2446.0000
#>    year
#> age 2005       2006      
#>   1    81.0000     0.0005
#>   2   291.0000     0.0005
#> 
#> units:  NA 

is(inits.ane)

#> [1] "FLPar"     "array"     "structure" "vector"   

inits.ane

#> An object of class "FLPar"
#> params
#>       q_depm   q_acoustic     psi_depm psi_acoustic      xi_depm  xi_acoustic 
#>            1            1           50           50            3            3 
#>           B0       R_1987       R_1988       R_1989       R_1990       R_1991 
#>        50000        80000        80000        80000        80000        80000 
#>       R_1992       R_1993       R_1994       R_1995       R_1996       R_1997 
#>        80000        80000        80000        80000        80000        80000 
#>       R_1998       R_1999       R_2000       R_2001       R_2002       R_2003 
#>        80000        80000        80000        80000        80000        80000 
#>       R_2004       R_2005       R_2006          mur         psir 
#>        80000        80000        80000           10            2 
#> units:  NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA 

is(control.ane)

#> [1] "bbmControl"

control.ane

#> An object of class "bbmControl"
#> Slot "g":
#>   rec adult 
#>  0.68  0.68 
#> 
#> Slot "param.fix":
#> An object of class "FLPar"
#> params
#>       q_depm   q_acoustic     psi_depm psi_acoustic      xi_depm  xi_acoustic 
#>            0            0            0            0            0            0 
#>           B0       R_1987       R_1988       R_1989       R_1990       R_1991 
#>            0            0            0            0            0            0 
#>       R_1992       R_1993       R_1994       R_1995       R_1996       R_1997 
#>            0            0            0            0            0            0 
#>       R_1998       R_1999       R_2000       R_2001       R_2002       R_2003 
#>            0            0            0            0            0            0 
#>       R_2004       R_2005       R_2006          mur         psir 
#>            0            0            0            0            0 
#> units:  NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA NA 
#> 

is(indicesB.ane)

#> [1] "FLIndices" "FLComps"   "FLlst"     "list"      "vector"   

indicesB.ane

#> An object of class "FLIndices"
#> [[1]]
#> An object of class "FLIndex"
#> 
#> Name:  
#> Description:  
#> Type :   
#> Distribution :   
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	1	20	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	startf	endf 
#> 	NA	NA	NA	1987	2006	0.375	0.375	
#> 
#> index         : [ 1 20 1 1 1 1 ], units =  NA 
#> index.var     : [ 1 20 1 1 1 1 ], units =  NA 
#> catch.n       : [ 1 20 1 1 1 1 ], units =  NA 
#> catch.wt      : [ 1 20 1 1 1 1 ], units =  NA 
#> effort        : [ 1 20 1 1 1 1 ], units =  NA 
#> sel.pattern   : [ 1 20 1 1 1 1 ], units =  NA 
#> index.q       : [ 1 20 1 1 1 1 ], units =  NA 
#> 
#> [[2]]
#> An object of class "FLIndex"
#> 
#> Name:  
#> Description:  
#> Type :   
#> Distribution :   
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	1	20	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	startf	endf 
#> 	NA	NA	NA	1987	2006	0.375	0.375	
#> 
#> index         : [ 1 20 1 1 1 1 ], units =  NA 
#> index.var     : [ 1 20 1 1 1 1 ], units =  NA 
#> catch.n       : [ 1 20 1 1 1 1 ], units =  NA 
#> catch.wt      : [ 1 20 1 1 1 1 ], units =  NA 
#> effort        : [ 1 20 1 1 1 1 ], units =  NA 
#> sel.pattern   : [ 1 20 1 1 1 1 ], units =  NA 
#> index.q       : [ 1 20 1 1 1 1 ], units =  NA 
#> 
#> Slot "names":
#> [1] "depm"     "acoustic"
#> 
#> Slot "desc":
#> character(0)
#> 
#> Slot "lock":
#> [1] FALSE
#> 

is(indicesP.ane)

#> [1] "FLIndices" "FLComps"   "FLlst"     "list"      "vector"   

indicesP.ane

#> An object of class "FLIndices"
#> [[1]]
#> An object of class "FLIndex"
#> 
#> Name:  
#> Description:  
#> Type :   
#> Distribution :   
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	1	20	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	startf	endf 
#> 	NA	NA	NA	1987	2006	0.375	0.375	
#> 
#> index         : [ 1 20 1 1 1 1 ], units =  NA 
#> index.var     : [ 1 20 1 1 1 1 ], units =  NA 
#> catch.n       : [ 1 20 1 1 1 1 ], units =  NA 
#> catch.wt      : [ 1 20 1 1 1 1 ], units =  NA 
#> effort        : [ 1 20 1 1 1 1 ], units =  NA 
#> sel.pattern   : [ 1 20 1 1 1 1 ], units =  NA 
#> index.q       : [ 1 20 1 1 1 1 ], units =  NA 
#> 
#> [[2]]
#> An object of class "FLIndex"
#> 
#> Name:  
#> Description:  
#> Type :   
#> Distribution :   
#> Quant: age 
#> Dims:  age 	year	unit	season	area	iter
#> 	1	20	1	1	1	1	
#> 
#> Range:  min	max	pgroup	minyear	maxyear	startf	endf 
#> 	NA	NA	NA	1987	2006	0.375	0.375	
#> 
#> index         : [ 1 20 1 1 1 1 ], units =  NA 
#> index.var     : [ 1 20 1 1 1 1 ], units =  NA 
#> catch.n       : [ 1 20 1 1 1 1 ], units =  NA 
#> catch.wt      : [ 1 20 1 1 1 1 ], units =  NA 
#> effort        : [ 1 20 1 1 1 1 ], units =  NA 
#> sel.pattern   : [ 1 20 1 1 1 1 ], units =  NA 
#> index.q       : [ 1 20 1 1 1 1 ], units =  NA 
#> 
#> Slot "names":
#> [1] "depm"     "acoustic"
#> 
#> Slot "desc":
#> character(0)
#> 
#> Slot "lock":
#> [1] FALSE
#>